Add files via upload
This is a Notification system made using python for Programmers. This Notification runs only in office hours(i.e. between 9 am to 5 pm). It will help in reminding programmers to drink water in every 20 minutes,  perform a eye exercise in every 30 minutes and to do a full body exercise in every 45 minutes, By doing this, programmer will remain healthy and active